Make `apply` instance methods public
Swift 2.0 lets us create public extensions on generic types again!
Huzzah! That means we can publicize `apply`, which will be nice for when
using an instance method makes more sense than using an operator.

This is something I always _wanted_ to do, but was never _able_ to do.
Pumped to make this public.