Improve consistency with generic naming
We should _always_ be using T, U, and V for generic type naming. We had some
places where we were still using A, B, and C, and a couple of places where we
only used U. Fixing this makes no difference to the use of the code, but is
more consistent internally.