Add support for building w/ Swift Package Manager
This does _not_ add support for running `swift test`, unfortunately. The test
dependency stuff seems to not be fully flushed out in the current build and
AFAIK, there's no way to point to a specific branch, only a tag. So we can't
actually point to the proper version of SwiftCheck right now anyway.