Remove swift-test support
Swift Package Manager doesn't hide test dependencies from transitive
dependencies. It _also_ doesn't support using branch-based dependencies for
stable releases. This means that if another library (Argo) depends on this
library, it will blow up when trying to build because it sees an incompatible
dependency (SwiftCheck pointed at a branch) even though that dependency is
_only_ used for testing.

This is super unfortunate, but there's no way around it currently. In order to
fix the ability to use Runes as a transitive dependency, we need to also
remove the dependency on SwiftCheck, which means we need to drop support for
`swift-test`.