Add support for Result
This adds support for _almost_ all of our operators for the standard lib's
`Result` type in Swift 5. I say almost because it's impossible to write
`empty()` for `Result<T, E>`, so I just... like... didn't.

Anyway, a cool thing to point out is that basically all of this is copy/pasted
from the Optional implementations (including the tests). All I had to do was
update some types and flesh out `Applicative` and we're in business.